\name{anaglyph}
\alias{anaglyph}
\title{Draw anaglyphs with base graphics.}
\usage{anaglyph(d3, blue, red)}

\description{
  Draw anaglyphs with base graphics.
}
\keyword{internal}
\arguments{
  \item{d3}{3d numeric matrix giving position of points}
  \item{blue}{blue colour (for right eye)}
  \item{red}{red colour (for left eye)}
}
