% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section-pursuit.r
\name{angular_breaks}
\alias{angular_breaks}
\title{Returns n equidistant bins between -pi and pi}
\usage{
angular_breaks(n)
}
\arguments{
\item{n}{number of bins}
}
\description{
Returns n equidistant bins between -pi and pi
}
