% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly-pursuit.r
\name{anomaly_index}
\alias{anomaly_index}
\title{Anomaly index.}
\usage{
anomaly_index()
}
\description{
Calculates an index that looks for the best projection of
observations that are outside a pre-determined p-D ellipse.
}
