% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bases_little}
\alias{bases_little}
\title{Generate bases for the little tour}
\usage{
bases_little(p, d = 2)
}
\arguments{
\item{d}{dimensionality of target projection}

\item{n}{dimensionality of data}
}
\description{
Generate bases for the little tour
}
\keyword{internal}

