% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour-dependence.r
\name{basis_dep_init}
\alias{basis_dep_init}
\title{Initial basis for dependence tour.}
\usage{
basis_dep_init(n, d, pos)
}
\arguments{
\item{n}{dimensionality of data}

\item{d}{dimensionality of target projection}

\item{pos}{a numeric vector describing which variables are mapped to
which dimensions: 1 corresponds to first, 2 to second etc.}
}
\description{
First two variables are projected on first two axes.
}
\keyword{internal}
