\name{basis_init}
\alias{basis_init}
\title{Generate initial basis.}
\usage{basis_init(n, d)}

\description{
  Generate initial basis.
}

\details{
  First two variables are projected on first two axes.
}
\keyword{internal}
\arguments{
  \item{n}{dimensionality of data}
  \item{d}{dimensionality of target projection}
}
