% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour-odd.R
\name{basis_init_odd}
\alias{basis_init_odd}
\title{Generate initial basis.}
\usage{
basis_init_odd(
  n = ncol(data),
  d = 2,
  data,
  n_points = n_points,
  n_hist = n_hist,
  ...
)
}
\arguments{
\item{n}{dimensionality of data}

\item{d}{dimensionality of target projection}
}
\description{
First two variables are projected on first two axes.
}
\keyword{internal}
