% Generated by roxygen2 (4.0.2): do not edit by hand
\name{basis_nearby}
\alias{basis_nearby}
\title{Generate nearby bases, e.g. for simulated annealing.}
\usage{
basis_nearby(current, alpha = 0.5, method = "linear")
}
\description{
Generate nearby bases, e.g. for simulated annealing.
}
\keyword{internal}

