% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour-grand.r
\name{basis_random}
\alias{basis_random}
\title{Generate a random basis}
\usage{
basis_random(n, d = 2)
}
\arguments{
\item{n}{dimensionality of data}

\item{d}{dimensionality of target projection}
}
\description{
Generate a random basis
}
\keyword{internal}
