% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour-odd.R
\name{basis_random_odd}
\alias{basis_random_odd}
\title{Generate a sample of random bases}
\usage{
basis_random_odd(
  n = ncol(data),
  d = 2,
  data,
  n_points = n_points,
  n_hist = n_hist
)
}
\arguments{
\item{n}{dimensionality of data}

\item{d}{dimensionality of target projection}

\item{n_points}{number of projections to generate}
}
\description{
Generate a sample of random bases
}
\keyword{internal}
