% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_index.R
\name{calc_index}
\alias{calc_index}
\title{calculate index value given the dataset and projection bases using holes()
created to allow calculate index outside the whole loop}
\usage{
calc_index(data, proj, sphere = TRUE)
}
\arguments{
\item{data}{the initial data used}

\item{proj}{the projection basis}

\item{sphere}{a binary indicator whether the data need to be sphered}
}
\description{
calculate index value given the dataset and projection bases using holes()
created to allow calculate index outside the whole loop
}
