% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{center}
\alias{center}
\title{Center a numeric vector by subtracting off its mean.}
\usage{
center(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Center a numeric vector by subtracting off its mean.
}
