% Generated by roxygen2 (4.0.0): do not edit by hand
\name{check_freezer_safe}
\alias{check_freezer_safe}
\title{Check matrix is a valid frozen matrix}
\usage{
check_freezer_safe(frozen)
}
\arguments{
\item{frozen}{matrix to check for freezability}
}
\description{
Check matrix is a valid frozen matrix
}
\keyword{internal}

