\name{cm}
\alias{cm}
\title{Central mass index.}
\usage{cm(mat)}

\description{
  Central mass index.
}

\details{
  Calculates the central mass index.  See Cook and Swayne (2007)
  Interactive and Dynamic Graphics for Data Analysis for equations.
}
\keyword{hplot}
\arguments{
  \item{mat}{matrix being used}
}
