% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cmass}
\alias{cmass}
\title{Central mass index.}
\usage{
cmass(mat)
}
\arguments{
\item{mat}{matrix being used}
}
\description{
Calculates the central mass index.  See Cook and Swayne (2007)
Interactive and Dynamic Graphics for Data Analysis for equations.
}
\keyword{hplot}

