% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicing-util.r
\name{compute_epsilon}
\alias{compute_epsilon}
\title{If not set, compute epsilon based on half_range}
\usage{
compute_epsilon(eps, half_range, n)
}
\description{
If not set, compute epsilon based on half_range
}
\keyword{internal}
