% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicing-util.r
\name{compute_v_rel}
\alias{compute_v_rel}
\title{If not set, compute epsilon based on half_range}
\usage{
compute_v_rel(v_rel, half_range, n)
}
\description{
If not set, compute epsilon based on half_range
}
\keyword{internal}
