% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-better.r
\name{correct_orientation}
\alias{correct_orientation}
\title{check if the current and target bases are of the same orientation}
\usage{
correct_orientation(current, target)
}
\description{
check if the current and target bases are of the same orientation
}
\keyword{internal}
