% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section-pursuit.r
\name{cumulative_radial}
\alias{cumulative_radial}
\title{Cumulative distribution in radius.}
\usage{
cumulative_radial(r, R, p)
}
\description{
Fraction of points within radius r given 2D projection of a
hypersphere with radius R in p dimensions.
}
\keyword{internal}
