\name{Kangaroo skull measurements}
\docType{data}
\alias{roos}
\title{Kangaroo skull measurements}
\description{
 
Data on kangaroos in australia, which comes from the
Andrews and Herzberg Data book.

The problem is that there are 3 historical skulls (housed in
different museums in europe) whose species identification got lost
somewhere along the line. so measurements were collected from 3
different species, both males and females from herds near where the
originals were collected in order to identify the historical skulls.
one complication is that kangaroo skulls keep growing throughout life
and they can be affected by diet. the historical skulls came from
kangaroos transported to europe in the late 1800s fed on a bread and
beer diet!

}
\usage{data(roos)}
\format{A 151 x 21 numeric array}
\keyword{datasets}
\examples{

}

