\name{SAT scores}
\docType{data}
\alias{sat}
\title{Sample SAT scores from 1988-89}
\description{
 
SAT Scores

Many administrators, teachers and parents perceive a "crisis in
education" in the USA. Learned skills and test scores suggest lower
achievement by students both over time and compared to their
counterparts around the world. Much of the concern is whether our
children will be able to compete in a global economy if they are
underprepared to meet the challenges of the coming decades.

Using the SAT scores of high school children as a measure of
educational achievement, does it appear that the level of financial
resources committed to schools in the U.S. influence educational
success? What does influence SAT scores of U.S. students?

The data contains 51 observations on the states and 10 variables for
the school year 1988-1989. The variables in order are:
\itemize{
  \item state, state, in alphabetical order
  \item sat, average combined verbal/math score (400-1600)
  \item expend, total expenditure for elementary and secondary public education used for instructional, noninstructional and support services in thousands of dollars.
  \item perpupil, total amount ($) per pupil
  \item enroll, enrollment, Fall 1988
  \item teachers, total number of teachers, measured in full-time equivalents
  \item ratio, pupil/Teacher ratio
  \item salary, average teacher salary
  \item attend, average daily attendance
  \item computer, record number in file
}

}
\usage{data(sat)}
\format{A 64 x 4 numeric array}
\keyword{datasets}
\examples{

}

