\name{Shipman patients}
\docType{data}
\alias{Shipman}
\title{Shimpman medical murder practice}
\description{
This data describes an investigation into the convicted murderer 
Dr. Harold Shipman. Shipman had lived in Hyde, England, it's been
determined that he killed at least 215 of his patients during his 
23 years of medical practice.

The records are ordered by time.

A number of cases (a few hundred?) were excluded from consideration:
patients of Shipman's who had remained alive and well for some time
after his last visit, for instance, or patients who died in hospital
after significant treatment, or those who died abroad or in
accidents.  So these are all cases where Shipman could not be
quickly excluded from consideration.

It is reported that one early clue was the clustering of deaths
at certain times of day, but the time of death isn't available
on the web site.

This material is protected by Crown copyright.
}
\usage{data(Shipman)}
\format{A 508 x 7 numeric array}
\keyword{datasets}
\references{http://www.the-shipman-inquiry.org.uk}
\examples{

}

