\name{Tropical Atmosphere Ocean}
\docType{data}
\alias{tao}
\title{Tropical Atmosphere Ocean data}
\description{

This is a subset of data taken from the NOAA web site
http://www.pmel.noaa.gov/tao/. The data is generated from recording
instruments on a grid of buoys laid out over the Pacific Ocean.  The
grid was setup to monitor El Nino and La Nina events. This subset
contains measurements from 5 locations (0deg/110W, 2S/110W,
0deg/95W,2S/95W,5S/95W) and two time points Nov-Jan 1993 (normal),
1997 (El Nino). There are missing values in this data set, which need
to be removed, or imputed before running a tour.

}
\usage{data(tao)}
\format{A 736 x 8 numeric array}
\references{http://www.pmel.noaa.gov/tao/}
\keyword{datasets}
\examples{
}

