\name{Tip data}
\docType{data}
\alias{tips}
\title{Tip data from Wallabys}
\description{
 
\itemize{
  \item obs, record number
  \item totbill, total of the bill for the table
  \item tip, tip given
  \item sex, sex of the payer
  \item smoker, the table in smoking section
  \item day, day served
  \item time, time served
  \item size, number of people served
}

}
\usage{data(tips)}
\format{A 244 x 8 numeric array}
\keyword{datasets}
\examples{

}
