% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interesting-indices.r
\name{dcor2d}
\alias{dcor2d}
\alias{dcor2d_2}
\title{Distance correlation index.}
\usage{
dcor2d()

dcor2d_2()
}
\description{
Computes the distance correlation based index on 2D projections of the data.
\code{dcor2d_2} uses the faster implementation of the distance correlation
for bivariate data, see \code{energy::dcor2d}.
}
\keyword{hplot}
