% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-dependence.r
\name{display_dependence}
\alias{display_dependence}
\alias{animate_dependence}
\title{Display tour path with two groups of variables}
\usage{
display_dependence(
  center = TRUE,
  axes = "center",
  half_range = NULL,
  col = "black",
  pch = 20,
  cex = 1,
  edges = NULL,
  edges.col = "black",
  edges.width = 1,
  obs_labels = NULL,
  fit = FALSE,
  palette = "Zissou 1",
  shapeset = c(15:17, 23:25),
  axislablong = FALSE,
  ...
)

animate_dependence(
  data,
  tour_path = dependence_tour(),
  display = display_dependence(),
  ...
)
}
\arguments{
\item{center}{if TRUE, centers projected data to (0,0).  This pins the
center of data cloud and make it easier to focus on the changing shape
rather than position.}

\item{axes}{position of the axes: center, bottomleft or off}

\item{half_range}{half range to use when calculating limits of projected.
If not set, defaults to maximum distance from origin to each row of data.}

\item{col}{color to use for points, can be a vector or hexcolors or a factor.  Defaults to "black".}

\item{pch}{shape of the point to be plotted, can be a factor or integer.  Defaults to 20.}

\item{cex}{size of the point to be plotted.  Defaults to 1.}

\item{edges}{A two column integer matrix giving indices of ends of lines.}

\item{edges.col}{colour of edges to be plotted, Defaults to "black"}

\item{edges.width}{line width for edges, default 1}

\item{obs_labels}{vector of text labels to display}

\item{fit}{logical indicating where to add a loess line onto plot}

\item{palette}{name of color palette for point colour, used by \code{\link{hcl.colors}}, default "Zissou 1"}

\item{shapeset}{numbers corresponding to shapes in base R points, to use for mapping
categorical variable to shapes, default=c(15:17, 23:25)}

\item{axislablong}{text labels only for the long axes in a projection, default FALSE}

\item{...}{other arguments passed on to \code{\link{animate}}}

\item{data}{matrix, or data frame containing numeric columns}

\item{tour_path}{tour path generator, defaults to 2d grand tour}

\item{display}{takes the display that is suppose to be used, defaults to
the xy display}
}
\description{
Animate a 2D tour path with a scatterplot.
}
\examples{
animate_dependence(flea[, 1:3],
  dependence_tour(c(1, 2, 2)),
  display_dependence(fit = TRUE, axes = "off"))

}
