% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-groupxy.r
\name{display_groupxy}
\alias{display_groupxy}
\alias{animate_groupxy}
\title{Display 2D tour projections displayed separately by groups}
\usage{
display_groupxy(center = TRUE, axes = "center", half_range = NULL,
  col = "black", pch = 20, edges = NULL, group_by = NULL,
  plot_gps = "all", plot_xgp = TRUE, gp_legend = TRUE, ...)

animate_groupxy(data, tour_path = grand_tour(), ...)
}
\arguments{
\item{center}{if TRUE, centers projected data to (0,0).  This pins the
center of data cloud and make it easier to focus on the changing shape
rather than position.}

\item{axes}{position of the axes: center, bottomleft or off}

\item{half_range}{half range to use when calculating limits of projected.
If not set, defaults to maximum distance from origin to each row of data.}

\item{col}{color to be plotted.  Defaults to "black".}

\item{pch}{shape of the point to be plotted.  Defaults to 20.}

\item{edges}{A two column integer matrix giving indices of ends of lines.}

\item{group_by}{variable to group by. Must have less than 25 unique values.}

\item{plot_gps}{vector of unique levels of the group_by variable to plot.}

\item{plot_xgp}{if TRUE, plots points from other groups in light grey.}

\item{gp_legend}{if TRUE, adds a legend of the group at the top left.}

\item{...}{other arguments passed on to \code{\link{animate}} and
\code{\link{display_groupxy}}}

\item{data}{matrix, or data frame containing numeric columns}

\item{tour_path}{tour path generator, defaults to 2d grand tour}
}
\description{
This function is designed to allow comparisons across multiple groups,
 especially for examining things like two (or more) different models
 on the same data. The primary display is a scatterplot, with lines or
 contours overlaid.
}
\examples{
animate_groupxy(flea[, 1:6])
animate(flea[, 1:6], tour_path=grand_tour(), display=display_groupxy())
animate(flea[, 1:6], tour_path=grand_tour(),
  display=display_groupxy(half_range = 0.5))
animate_groupxy(flea[, 1:6], tour_path=little_tour())
animate_groupxy(flea[, 1:3], tour_path=guided_tour(holes), sphere = TRUE)
animate_groupxy(flea[, 1:6], center = FALSE)

# The default axes are centered, like a biplot, but there are other options
animate_groupxy(flea[, 1:6], axes = "off")
animate_groupxy(flea[, 1:6], dependence_tour(c(1, 2, 1, 2, 1, 2)) )

# Adding color and group
require(colorspace)
pal <- rainbow_hcl(length(levels(flea$species)))
col <- pal[as.numeric(flea$species)]
gp = flea$species
animate_groupxy(flea[,-7], col=col, group_by=gp)

# You can also draw lines
edges <- matrix(c(1:5, 2:6), ncol = 2)
animate(flea[, 1:6], grand_tour(),
  display_groupxy(axes="bottomleft", edges=edges, group_by=gp, col=col))

# Point characters, plotting other groups, and plotting legends
pch <- as.numeric(flea$species)+14
animate_groupxy(flea[, 1:6], col = col, pch = pch, group_by = gp)
animate_groupxy(flea[, 1:6], col = col, pch = pch, group_by = gp, plot_xgp = FALSE)
animate_groupxy(flea[, 1:6], col = col, pch = pch, group_by = gp, gp_legend = FALSE)
}
