% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-idx.R
\name{display_idx}
\alias{display_idx}
\alias{animate_idx}
\title{Display 1D linear aggregation index}
\usage{
display_idx(
  center = FALSE,
  half_range = NULL,
  abb_vars = TRUE,
  col = "red",
  cex = 0.5,
  label_x_pos = 0.7,
  label = NULL,
  label_cex = 1,
  label_col = "grey80",
  palette = "Zissou 1",
  ...
)

animate_idx(data, tour_path = grand_tour(1), ...)
}
\arguments{
\item{center}{should 1d projection be centered to have mean zero (default: TRUE).
This pins the centre of distribution to the same place, and makes it
easier to focus on the shape of the distribution.}

\item{half_range}{half range to use when calculating limits of projected.
If not set, defaults to maximum distance from origin to each row of data.}

\item{col}{the color used for points, can be a vector or hexcolors or a
factor, default to "red".}

\item{cex}{the size used for points, default to 0.5}

\item{label_x_pos}{the x position of text label, currently labels are
positioned at a fixed x value for each observation}

\item{label}{the text label, a vector}

\item{label_cex}{the size for text labels}

\item{label_col}{the color for text labels}

\item{palette}{name of color palette for point colour, used by
\code{\link{hcl.colors}}, default "Zissou 1"}

\item{...}{ignored}

\item{data}{matrix, or data frame containing numeric columns}

\item{tour_path}{tour path generator, defaults to 2d grand tour}
}
\description{
Display 1D linear aggregation index
}
