% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-slice-bin.r
\name{display_slice_bin}
\alias{display_slice_bin}
\title{Display tour path with a sliced scatterplot}
\usage{
display_slice_bin(
  center = TRUE,
  axes = "center",
  half_range = NULL,
  col = "black",
  pch_slice = 20,
  pch_other = 46,
  eps = NULL,
  anchor = NULL,
  edges = NULL,
  nbin = 30,
  edges.col = "black",
  ...
)
}
\arguments{
\item{center}{if TRUE, centers projected data to (0,0).  This pins the
center of data cloud and make it easier to focus on the changing shape
rather than position.}

\item{axes}{position of the axes: center, bottomleft or off}

\item{half_range}{half range to use when calculating limits of projected.
If not set, defaults to maximum distance from origin to each row of data.}

\item{col}{color to be plotted.  Defaults to "black"}

\item{pch_slice}{marker for plotting points inside the slice.
Defaults to 20.}

\item{pch_other}{marker for plotting points outside the slice.
Defaults to 46.}

\item{eps}{volume of the slice. If not set, suggested value is caluclated and
printed to the screen.}

\item{anchor}{A vector specifying the reference point to anchor the slice.
If NULL (default) the slice will be anchored at the origin.}

\item{edges}{A two column integer matrix giving indices of ends of lines.}

\item{edges.col}{colour of edges to be plotted, Defaults to "black.}

\item{...}{other arguments passed on to \code{\link{animate}} and
\code{\link{display_slice}}}

\item{rescale}{if true, rescale all variables to range [0,1].}
}
\description{
Animate a 2D tour path with a sliced scatterplot.
}
\examples{
# Generate samples on a 3d and 5d hollow sphere using the geozoo package
sphere3 <- geozoo::sphere.hollow(3)$points
sphere5 <- geozoo::sphere.hollow(5)$points

# Columns need to be named before launching the tour
colnames(sphere3) <- c("x1", "x2", "x3")
colnames(sphere5) <- c("x1", "x2", "x3", "x4", "x5")

# Animate with the slice display using the default parameters
animate_slice(sphere3)
animate_slice(sphere5)

# Animate with off-center anchoring
anchor3 <- rep(0.7, 3)
anchor5 <- rep(0.3, 5)
animate_slice(sphere3, anchor = anchor3)
# Animate with thicker slice to capture more points in each view
animate_slice(sphere5, anchor = anchor5, eps = 0.02)
}
