% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-slice.r
\name{draw_slice_center}
\alias{draw_slice_center}
\title{Draw slice center guide with base graphics}
\usage{
draw_slice_center(anchor, rng, limits, anchor_nav, ...)
}
\description{
Draw slice center guide with base graphics
}
\keyword{internal}
