% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-xy.r
\name{draw_tour_axes}
\alias{draw_tour_axes}
\title{Draw tour axes with base graphics}
\usage{
draw_tour_axes(proj, labels, limits, position)
}
\description{
Draw tour axes with base graphics
}
\keyword{internal}
