% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicing-util.r
\name{eucl_norm_sq}
\alias{eucl_norm_sq}
\title{Calculate the squared Euclidean norm of a vector x}
\usage{
eucl_norm_sq(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Calculate the squared Euclidean norm of a vector x
}
\keyword{internal}
