% Generated by roxygen2 (4.0.0): do not edit by hand
\name{find_frozen_path_peak}
\alias{find_frozen_path_peak}
\title{Find most highest peak along frozen geodesic.}
\usage{
find_frozen_path_peak(old, new, frozen, index, max_dist = pi/4)
}
\description{
Find most highest peak along frozen geodesic.
}
\keyword{internal}

