% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{find_platform}
\alias{find_platform}
\title{Find the platform
Find the platform being used by the user}
\usage{
find_platform()
}
\description{
Find the platform
Find the platform being used by the user
}
\keyword{internal}
