% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour-guided-dep.R
\name{guided_tour_dep}
\alias{guided_tour_dep}
\title{A guided tour path for}
\usage{
guided_tour_dep(
  index_f,
  pos,
  d = 2,
  alpha = 0.5,
  cooling = 0.99,
  max.tries = 25,
  max.i = Inf,
  search_f = search_geodesic,
  n_sample = 100,
  ...
)
}
\description{
A guided tour path for
}
