\name{holes}
\alias{holes}
\title{Holes index.}
\usage{holes(mat)
holes(mat)
holes(mat)
holes(mat)
holes(mat)
}
\description{Holes index.}
\details{Calculates the holes index. See Cook and Swayne (2007)
Interactive and Dynamic Graphics for Data Analysis for equations.\code{holes}: Calculates the holes index. See Cook and Swayne (2007)
Interactive and Dynamic Graphics for Data Analysis for equations.


\code{holes}: Calculates the holes index. See Cook and Swayne (2007)
Interactive and Dynamic Graphics for Data Analysis for equations.


\code{holes}: Calculates the holes index. See Cook and Swayne (2007)
Interactive and Dynamic Graphics for Data Analysis for equations.


\code{holes}: Calculates the holes index. See Cook and Swayne (2007)
Interactive and Dynamic Graphics for Data Analysis for equations.


}
\keyword{hplot}
\arguments{\item{mat}{matrix being used}
}
\alias{holes}
\keyword{hplot}

\alias{holes}
\keyword{hplot}

\alias{holes}
\keyword{hplot}

\alias{holes}
\keyword{hplot}

