% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interesting-indices.r
\name{MIC}
\alias{MIC}
\alias{TIC}
\title{Maximum and total information coefficient index.}
\usage{
MIC()

TIC()
}
\description{
Compute the maximum and total information coefficient indexes,
see \code{minerva::mine}.
}
