\name{lda_pp}
\alias{lda_pp}
\title{LDA projection pursuit index.}
\usage{lda_pp(cl)}

\description{
  LDA projection pursuit index.
}

\details{
  Calculate the LDA projection pursuit index.  See Cook and
  Swayne (2007) Interactive and Dynamic Graphics for Data
  Analysis for equations.
}
\keyword{hplot}
\arguments{
  \item{cl}{class to be used.  Such as "color"}
}
