% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section-pursuit.r
\name{linear_breaks}
\alias{linear_breaks}
\title{Returns n equidistant bins between a and b}
\usage{
linear_breaks(n, a, b)
}
\arguments{
\item{n}{number of bins}

\item{a}{lower bound}

\item{b}{upper bound}
}
\description{
Returns n equidistant bins between a and b
}
