% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-algebra.r
\name{mahal_dist}
\alias{mahal_dist}
\title{Calculate the Mahalanobis distance between points and center.}
\usage{
mahal_dist(x, vc)
}
\arguments{
\item{x}{matrix of data}

\item{vc}{pre-determined variance-covariance matrix}
}
\description{
Computes the Mahalanobis distance using a provided variance-covariance
matrix of observations from 0.
}
\keyword{algebra}
