% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{mapColors}
\alias{mapColors}
\title{Map vector of factors to color}
\usage{
mapColors(x, palette)
}
\arguments{
\item{x}{vector}

\item{palette}{name of color palette for point colour, used by \code{\link{hcl.colors}}, default "Zissou 1"}
}
\description{
Map vector of factors to color
}
