% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{mapShapes}
\alias{mapShapes}
\title{Map vector of factors to pch}
\usage{
mapShapes(x, shapeset)
}
\arguments{
\item{x}{vector}

\item{shapeset}{vector of integers indicating point shapes}
}
\description{
Map vector of factors to pch
}
