% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{norm}
\alias{norm}
\title{A generated random sequence}
\format{
A vector of length 1000
}
\description{
This sequence is used in norm_bin() and norm_kol() to ensure the index for
a fixed projection is always the same
}
