% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interesting-indices.r
\name{norm_bin}
\alias{norm_bin}
\alias{norm_kol}
\title{Normality index.}
\usage{
norm_bin()

norm_kol()
}
\description{
Compares the similarity between the projected distribution and a normal distribution.
\itemize{
\item{norm_bin }{compares the count in 100 histogram bins}
\item{norm_kol }{compares the cdf based on the Kolmogorov–Smirnov test (KS test)}
}
}
\keyword{hplot}
