\name{normalise}
\alias{normalise}
\title{Normalise a numeric matrix.}
\usage{normalise(x)
normalise(x)
normalise(x)
normalise(x)
normalise(x)
}
\description{Normalise a numeric matrix.}
\details{Ensure that columns of a numeric matrix have norm 1\code{normalise}: Ensure that columns of a numeric matrix have norm 1


\code{normalise}: Ensure that columns of a numeric matrix have norm 1


\code{normalise}: Ensure that columns of a numeric matrix have norm 1


\code{normalise}: Ensure that columns of a numeric matrix have norm 1


}
\keyword{internal}
\keyword{algebra}
\arguments{\item{x}{numeric matrix or vector}
}
\alias{normalise}
\keyword{internal}
\keyword{algebra}

\alias{normalise}
\keyword{internal}
\keyword{algebra}

\alias{normalise}
\keyword{internal}
\keyword{algebra}

\alias{normalise}
\keyword{internal}
\keyword{algebra}

