\name{observe_vectors_move}
\alias{observe_vectors_move}
\title{Di Function...}
\usage{observe_vectors_move(basis_set, index_f, ...)}
\description{Di Function}
\keyword{internal}
\examples{#t1 <- save_history(flea[, c(1,2,4)], nbases = 3, d = 1)
#t1 <- save_history(flea[, 1:6], nbases = 3, d = 1)
#observe_vectors_move(t1)
#observe_vectors_move(interpolate(t1))}
