% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour-odd.R
\name{odd_tour}
\alias{odd_tour}
\title{An outlier tour path.}
\usage{
odd_tour(d = 2, n_points = 10, n_hist = 3, ...)
}
\arguments{
\item{d}{target dimensionality}

\item{...}{arguments sent to the generator}
}
\description{
This method generates a sample of target bases by randomly sampling
on the space of all d-dimensional planes in p-space and selecting
the basis most likely to be an outlier.
}
\details{
Usually, you will not call this function directly, but will pass it to
a method that works with tour paths like \code{\link{animate}},
\code{\link{save_history}} or \code{\link{render}}.
}
\examples{
# All animation methods use the grand tour path by default
animate_xy(flea[, 1:6], odd_tour(n_points = 5), col = flea$species)
}
