\name{orthonormalise}
\alias{orthonormalise}
\title{Orthonormalise using modified Gram-Schmidt process.}
\usage{orthonormalise(x)
orthonormalise(x)
orthonormalise(x)
orthonormalise(x)
orthonormalise(x)
}
\description{Orthonormalise using modified Gram-Schmidt process.}
\keyword{internal}
\keyword{algebra}
\arguments{\item{x}{numeric matrix}
}
\alias{orthonormalise}
\keyword{internal}
\keyword{algebra}

\alias{orthonormalise}
\keyword{internal}
\keyword{algebra}

\alias{orthonormalise}
\keyword{internal}
\keyword{algebra}

\alias{orthonormalise}
\keyword{internal}
\keyword{algebra}

