\name{orthonormalise_by}
\alias{orthonormalise_by}
\title{Orthonnormalise one matrix by another.}
\usage{orthonormalise_by(x, by)
orthonormalise_by(x, by)
orthonormalise_by(x, by)
orthonormalise_by(x, by)
orthonormalise_by(x, by)
}
\description{Orthonnormalise one matrix by another.}
\details{This ensures that each column in x is orthogonal to the corresponding
column in by.\code{orthonormalise_by}: This ensures that each column in x is orthogonal to the corresponding
column in by.


\code{orthonormalise_by}: This ensures that each column in x is orthogonal to the corresponding
column in by.


\code{orthonormalise_by}: This ensures that each column in x is orthogonal to the corresponding
column in by.


\code{orthonormalise_by}: This ensures that each column in x is orthogonal to the corresponding
column in by.


}
\keyword{internal}
\keyword{algebra}
\arguments{\item{x}{numeric matrix}
\item{by}{numeric matrix, same size as x}
}
\alias{orthonormalise_by}
\keyword{internal}
\keyword{algebra}

\alias{orthonormalise_by}
\keyword{internal}
\keyword{algebra}

\alias{orthonormalise_by}
\keyword{internal}
\keyword{algebra}

\alias{orthonormalise_by}
\keyword{internal}
\keyword{algebra}

