% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Ozone measurements}
\alias{Ozone measurements}
\alias{ozone}
\title{Monthly ozone measurements over Central America}
\format{
A 24 x 24 x 72 numeric array
}
\description{
This data set is a subset of the data from the 2006 ASA Data expo challenge,
\url{http://stat-computing.org/dataexpo/2006/}. The data are monthly ozone
averages on a very coarse 24 by 24 grid covering Central America, from Jan
1995 to Dec 2000. The data is stored in a 3d area with the first two
dimensions representing latitude and longitude, and the third representing
time.
}
\examples{

example(display_image)

}
\references{
\url{http://stat-computing.org/dataexpo/2006/}
}
\keyword{datasets}
