\name{pda_pp}
\alias{pda_pp}
\title{PDA projection pursuit index.}
\usage{pda_pp(cl, lambda)}
\description{PDA projection pursuit index.}
\details{Calculates the PDA projections pursuit index}
\keyword{hplot}
\arguments{\item{cl}{class to be used.  Such as "color"}
\item{lambda}{parameter used the "interesting" persuit.  It ranges from 0 to 1. O is lda_pp.  pda_pp is an updated form of lda_pp.}}
