\name{plot.history_curve}
\alias{plot.history_curve}
\title{Plot history curves.}
\usage{\method{plot}{history_curve} (x, ...)}
\description{Plot history curves.}
\details{The default plot method is a line plot with step on the x axis and 
value on the y axis.  Each observation is drawn with a different line
line and the plot is facetted by variable.  This is rather similar in 
spirit to a parallel coordinates plot or Andrews curves.

For alternative ways of plotting this data, see
\code{\link{history_curves}}}
\keyword{internal}
