\name{plot.path_index}
\alias{plot.path_index}
\title{Plot history index with ggplot2.}
\usage{\method{plot}{path_index}(x, ...)
\method{plot}{path_index}(x, ...)
\method{plot}{path_index}(x, ...)
\method{plot}{path_index}(x, ...)
\method{plot}{path_index}(x, ...)
}
\description{Plot history index with ggplot2.}
\keyword{internal}
\keyword{hplot}

\alias{plot.path_index}
\keyword{internal}
\keyword{hplot}

\alias{plot.path_index}
\keyword{internal}
\keyword{hplot}

\alias{plot.path_index}
\keyword{internal}
\keyword{hplot}

\alias{plot.path_index}
\keyword{internal}
\keyword{hplot}

