% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{print_final_proj}
\alias{print_final_proj}
\title{Print out the final projection basis}
\usage{
print_final_proj(basis)
}
\arguments{
\item{basis}{the projection basis to print, used in search_* functions}
}
\description{
Print out the final projection basis
}
\keyword{internal}
