\name{proj_dist}
\alias{proj_dist}
\title{Calculate the distance between two bases.}
\usage{proj_dist(x, y)
proj_dist(x, y)
proj_dist(x, y)
proj_dist(x, y)
proj_dist(x, y)
}
\description{Calculate the distance between two bases.}
\details{Computes the Frobenius norm between two bases.  This is equals to the
Euclidean norm of the vector of sines of principal angles between the two
subspaces.\code{proj_dist}: Computes the Frobenius norm between two bases.  This is equals to the
Euclidean norm of the vector of sines of principal angles between the two
subspaces.


\code{proj_dist}: Computes the Frobenius norm between two bases.  This is equals to the
Euclidean norm of the vector of sines of principal angles between the two
subspaces.


\code{proj_dist}: Computes the Frobenius norm between two bases.  This is equals to the
Euclidean norm of the vector of sines of principal angles between the two
subspaces.


\code{proj_dist}: Computes the Frobenius norm between two bases.  This is equals to the
Euclidean norm of the vector of sines of principal angles between the two
subspaces.


}
\keyword{algebra}
\arguments{\item{x}{projection matrix a}
\item{y}{projection matrix b}
}

\alias{proj_dist}
\keyword{algebra}

\alias{proj_dist}
\keyword{algebra}

\alias{proj_dist}
\keyword{algebra}

\alias{proj_dist}
\keyword{algebra}

