% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour-radial.r
\name{radial_tour}
\alias{radial_tour}
\title{A radial tour path.}
\usage{
radial_tour(start, var = 1)
}
\arguments{
\item{start}{initial projection matrix}

\item{var}{variable chosen to rotate out}
}
\description{
The radial tour rotates a chosen variable axis out of the current
projection.
}
\details{
Usually, you will not call this function directly, but will pass it to
a method that works with tour paths like \code{\link{animate}},
\code{\link{save_history}} or \code{\link{render}}.
}
\examples{
animate_xy(flea[, 1:6], radial_tour(basis_random(6, 2), var = 4))
animate_dist(flea[, 1:6], radial_tour(basis_random(6, 1), var = 4))
animate_scatmat(flea[, 1:6], radial_tour(basis_random(6, 3), var = 4))
}
