\name{rescale}
\alias{rescale}
\title{Rescale a matrix or data frame...}
\usage{rescale(df)
rescale(df)
}
\description{Rescale a matrix or data frame}
\details{Standardise each column to have range [0, 1]\code{rescale}: Standardise each column to have range [0, 1]


}
\keyword{manip}
\arguments{\item{df}{data frame or matrix}
}

\alias{rescale}
\keyword{manip}

