\name{search_better}
\alias{search_better}
\title{Search for a better projection near the current projection.}
\usage{
  search_better(current, alpha = 0.5, index, max.tries =
  Inf, method = "linear")
}
\description{
  Search for a better projection near the current
  projection.
}
\keyword{internal}

