% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-frozen-geodesic.r
\name{search_frozen_geodesic}
\alias{search_frozen_geodesic}
\title{A pseudo-derivative, line search algorithm along frozen geodesics.}
\usage{
search_frozen_geodesic(
  current,
  index,
  max.tries = 5,
  n = 5,
  frozen,
  cur_index = NA,
  ...
)
}
\description{
These three functions perform a corresponding role to
\code{\link{search_geodesic}}, \code{\link{find_best_dir}} and
\code{\link{find_path_peak}} but for the frozen tour.  They work by
zero'ing out the frozen variables and travelling in that restricted
subspace.
}
\section{To do}{
 eliminate these functions
}

\keyword{internal}
