% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-jellyfish.R
\name{search_jellyfish}
\alias{search_jellyfish}
\title{An jellyfish optimisers for the projection pursuit guided tour}
\usage{
search_jellyfish(current, index, tries, max.tries = 50, ...)
}
\arguments{
\item{current}{starting projection, a list of basis of class "multi-bases"}

\item{index}{index function}

\item{tries}{the counter of the outer loop of the opotimiser}

\item{max.tries}{the maximum number of iteration before giving up}

\item{...}{other arguments being passed into the \code{search_jellyfish()}}
}
\description{
An jellyfish optimisers for the projection pursuit guided tour
}
\examples{
res <- animate_xy(flea[, 1:6], guided_tour(holes(), search_f = search_jellyfish))
res
}
\keyword{optimize}
