% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_polish.R
\name{search_polish}
\alias{search_polish}
\title{Search for most interesting projection along random geodesics.}
\usage{
search_polish(
  current,
  polish_alpha = 0.05,
  index,
  max.tries = 5,
  cur_index = NA,
  n_sample = 1000,
  polish_cooling = 1,
  ...
)
}
\arguments{
\item{current}{the current projeciton basis}

\item{polish_alpha}{the angle used to search the target basis from the current basis}

\item{index}{index function}

\item{max.tries}{maximum number of iteration before giving up}

\item{cur_index}{the index value of the current basis}

\item{n_sample}{number of samples to generate}

\item{polish_cooling}{percentage of reductio in polish_alpha when no better basis is found}
}
\description{
Search for most interesting projection along random geodesics.
}
\keyword{You}
\keyword{\code{\link{guided_tour}}}
\keyword{a}
\keyword{as}
\keyword{but}
\keyword{call}
\keyword{directly,}
\keyword{function}
\keyword{have}
\keyword{it}
\keyword{not}
\keyword{optimize}
\keyword{search}
\keyword{should}
\keyword{strategy.}
\keyword{supply}
\keyword{the}
\keyword{this}
\keyword{to}
