% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_posse.R
\name{search_posse}
\alias{search_posse}
\title{Search for a better projection based on Poss, 1995}
\usage{
search_posse(
  current,
  alpha = 0.5,
  index,
  max.tries = 300,
  cur_index = NA,
  cooling = 0.9
)
}
\description{
Search for a better projection based on Poss, 1995
}
\keyword{internal}
