% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interesting-indices.r
\name{skewness}
\alias{skewness}
\title{Skewness index.}
\usage{
skewness()
}
\description{
Calculates the skewness index. See Cook, Buja and Cabrera (1993)
Projection pursuit indexes based on orthonormal function expansions
 for equations.
}
\keyword{hplot}
