% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interesting-indices.r
\name{skinny}
\alias{skinny}
\title{Scagnostic index: skinny}
\usage{
skinny()
}
\description{
Compute the skinny scagnostic measure from the `cassowaryr` package.
This index measures how thin and narrow a shape is.
}
