% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section-pursuit.r
\name{slice_binning}
\alias{slice_binning}
\title{Separately binning observations inside and outside the slice.}
\usage{
slice_binning(mat, dists, h, breaks_x, breaks_y, bintype = "square")
}
\arguments{
\item{mat}{projected data points}

\item{dists}{vector of all point distances from the projection plane.}

\item{h}{slice thickness.}

\item{breaks_x}{binning on the first variable (x or radius).}

\item{breaks_y}{binning on the second variable (y or angle).}

\item{bintype}{select polar (default) or square binning.}
}
\description{
Separately binning observations inside and outside the slice.
}
\keyword{internal}
