% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section-pursuit.r
\name{slice_index}
\alias{slice_index}
\title{Section pursuit index.}
\usage{
slice_index(
  breaks_x,
  breaks_y,
  eps,
  bintype = "polar",
  power = 1,
  flip = 1,
  reweight = FALSE,
  p = 4
)
}
\arguments{
\item{breaks_x}{binning on the first variable (x or radius).}

\item{breaks_y}{binning on the second variable (y or angle).}

\item{eps}{cutoff values to suppress summing up small differences.
Vector with one entry for each bin, can be estimated
using \code{\link{estimate_eps}}.}

\item{bintype}{select polar (default) or square binning.}

\item{power}{exponent q used in the index compuatation.}

\item{flip}{sign of the index computation, select +1 when searching
for low densities and -1 when searching for high densities.}

\item{reweight}{if TRUE will reweight according to the expected distribution
in a uniform hypersphere (default is FALSE).}

\item{p}{number of variables in the data (needed for accurate reweighting,
default is 4).}
}
\description{
Calculates a section pursuit index that compares the distribution
inside and outside a slice.
}
