\name{sphere}
\alias{sphere}
\title{Sphere a matrix (or data frame)...}
\usage{sphere(df)}
\description{Sphere a matrix (or data frame)}
\details{Sphering is often useful in conjunction with the guided tour, as it 
removes simpler patterns that may conceal more interesting findings.}
\keyword{manip}
\arguments{\item{df}{data frame or matrix}}
