% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{sphere_data}
\alias{sphere_data}
\title{Sphere a matrix (or data frame) by transforming variables to
principal components.}
\usage{
sphere_data(df)
}
\arguments{
\item{df}{data frame or matrix}
}
\description{
Sphering is often useful in conjunction with the guided tour, as it
removes simpler patterns that may conceal more interesting findings.
}
\keyword{manip}
