% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interesting-indices.r
\name{splines2d}
\alias{splines2d}
\title{Spline based index.}
\usage{
splines2d()
}
\description{
Compares the variance in residuals of a fitted
spline model to the overall variance to find
functional dependence in 2D projections
of the data.
}
\keyword{hplot}
