% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{standdise}
\alias{standdise}
\title{Standardise a matrix or data frame}
\usage{
standdise(df)
}
\arguments{
\item{df}{data frame or matrix}
}
\description{
Standardise each column to have mean 0 and var 1
}
\keyword{manip}
