% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{t1}
\alias{t1}
\title{Saved history of guided tour with holes}
\description{
This data was generated from the following code:
set.seed(2020)
t1 <- save_history(flea[, 1:6], guided_tour(holes()), max = 100)
attr(t1, "class") <- NULL
And used as an example for search_polish() to start
optimising from the best projection from search_geodesic.
t1 is a 3D array or 2D projections.
}
\keyword{datasets}
