% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{to_stop}
\alias{to_stop}
\title{Prints information on how to stop the output}
\usage{
to_stop()
}
\description{
This function prints the corresponding information on how to stop the plotting.
}
\keyword{internal}
