% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourr-package.R
\docType{package}
\name{tourr-package}
\alias{tourr}
\alias{tourr-package}
\title{tourr: Tour Methods for Multivariate Data Visualisation}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Implements geodesic interpolation and basis generation functions that allow you to create new tour methods from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggobi/tourr}
  \item Report bugs at \url{https://github.com/ggobi/tourr/issues}
}

}
\author{
\strong{Maintainer}: Di Cook \email{dicook@monash.edu} (\href{https://orcid.org/0000-0002-3813-7155}{ORCID})

Authors:
\itemize{
  \item Hadley Wickham \email{h.wickham@gmail.com} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item Nick Spyrison \email{nicholas.spyrison@monash.edu} [contributor]
  \item Ursula Laa \email{ursula.laa@boku.ac.at} (\href{https://orcid.org/0000-0002-0249-6439}{ORCID}) [contributor]
  \item H. Sherry Zhang \email{huizezhangsh@gmail.com} [contributor]
  \item Stuart Lee \email{stuart.lee1@monash.edu} (\href{https://orcid.org/0000-0003-1179-8436}{ORCID}) [contributor]
}

}
\keyword{internal}
