% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Tropical Atmosphere Ocean}
\alias{Tropical Atmosphere Ocean}
\alias{tao}
\title{Tropical Atmosphere Ocean data}
\format{
A 736 x 8 numeric array
}
\description{
This is a subset of data taken from the NOAA web site
http://www.pmel.noaa.gov/tao/. The data is generated from recording
instruments on a grid of buoys laid out over the Pacific Ocean.  The grid
was setup to monitor El Nino and La Nina events. This subset contains
measurements from 5 locations (0deg/110W, 2S/110W, 0deg/95W,2S/95W,5S/95W)
and two time points Nov-Jan 1993 (normal), 1997 (El Nino). There are missing
values in this data set, which need to be removed, or imputed before running
a tour.
}
\references{
http://www.pmel.noaa.gov/tao/
}
\keyword{datasets}
