\name{xnul}
\alias{xnul}
\alias{xnul}
\title{A null function...}
\usage{nul(...)
nul(...)
}
\description{A null function}
\details{This function does nothing, and is a useful default callback function\code{xnul}: This function does nothing, and is a useful default callback function


}
\keyword{internal}
\arguments{\item{...}{all arguments to \code{...} are ignore}
}

\alias{xnul}
\alias{xnul}
\keyword{internal}

