% Generated by roxygen2 (4.0.2): do not edit by hand
\name{anaglyph}
\alias{anaglyph}
\title{Draw anaglyphs with base graphics.}
\usage{
anaglyph(d3, blue, red)
}
\arguments{
\item{d3}{3d numeric matrix giving position of points}

\item{blue}{blue colour (for right eye)}

\item{red}{red colour (for left eye)}
}
\description{
Draw anaglyphs with base graphics.
}
\keyword{internal}

