% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour-little.r
\name{bases_little}
\alias{bases_little}
\title{Generate bases for the little tour}
\usage{
bases_little(p, d = 2)
}
\arguments{
\item{p}{dimensionality of data}

\item{d}{dimensionality of target projection}
}
\description{
Generate bases for the little tour
}
\keyword{internal}
