% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour-grand.r
\name{basis_init}
\alias{basis_init}
\title{Generate initial basis.}
\usage{
basis_init(n, d)
}
\arguments{
\item{n}{dimensionality of data}

\item{d}{dimensionality of target projection}
}
\description{
First two variables are projected on first two axes.
}
\keyword{internal}
