% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-better.r
\name{basis_nearby}
\alias{basis_nearby}
\title{Generate nearby bases, e.g. for simulated annealing.}
\usage{
basis_nearby(current, alpha = 0.5, method = "linear")
}
\description{
Generate nearby bases, e.g. for simulated annealing.
}
\keyword{internal}
