% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tour-frozen.r
\name{check_freezer_safe}
\alias{check_freezer_safe}
\title{Check matrix is a valid frozen matrix}
\usage{
check_freezer_safe(frozen)
}
\arguments{
\item{frozen}{matrix to check for freezability}
}
\description{
Check matrix is a valid frozen matrix
}
\keyword{internal}
