\name{cm}
\alias{cm}
\title{Central mass index.}
\usage{
  cm(mat)
}
\arguments{
  \item{mat}{matrix being used}
}
\description{
  Central mass index.
}
\details{
  Calculates the central mass index.  See Cook and Swayne
  (2007) Interactive and Dynamic Graphics for Data Analysis
  for equations.
}
\keyword{hplot}

