% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interesting-indices.r
\name{cmass}
\alias{cmass}
\title{Central mass index.}
\usage{
cmass()
}
\description{
Calculates the central mass index.  See Cook and Swayne (2007)
Interactive and Dynamic Graphics for Data Analysis for equations.
}
\keyword{hplot}
