% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-sage.R, R/section-pursuit.r
\name{cumulative_radial}
\alias{cumulative_radial}
\title{Calculate radial cumulative distribtuion}
\usage{
cumulative_radial(r, R, p)

cumulative_radial(r, R, p)
}
\arguments{
\item{r}{projected radius}

\item{R}{radius of pD sphere}

\item{p}{input dimension}
}
\value{
fraction of volume
}
\description{
Calculate fraction of volume of a pD sphere with radius R projected
within the projected radius r.

Fraction of points within radius r given 2D projection of a
hypersphere with radius R in p dimensions.
}
\keyword{internal}
