\name{Olive oil measurements}
\docType{data}
\alias{olive}
\title{Olive oil samples from Italy}
\description{
 
572 olive oil samples from nine different regions of Italy; for each sample the normalized concentrations of eight fatty acids are given. We thank Prof. Michele Forina, University of Genova, Italy for making this dataset available.

}
\usage{data(olive)}
\format{A 572 x 10 numeric array}
\keyword{datasets}
\examples{
}