\name{Tip data}
\docType{data}
\alias{tips}
\title{Tip data}
\description{This data is from Bryant and Smith (1995) Practical Data
Analysis: Case Studies in business statistics. It contains data
recorded by one waiter in one restaurant for a period of 2.5 months.
 
\itemize{
  \item obs, record number
  \item totbill, total of the bill for the table
  \item tip, tip given
  \item sex, sex of the payer
  \item smoker, the table in smoking section
  \item day, day served
  \item time, time served
  \item size, number of people served
}

}
\usage{data(tips)}
\format{A 244 x 8 numeric array}
\keyword{datasets}
\examples{

}
