\name{display_depth}
\title{Display 3d projection with depth cues...}
\usage{display_depth(...)}

\description{
  Display 3d projection with depth cues
}

\details{
  Suggestion to use gray background and colour saturation (instead of
  gray shading) by Graham Wills. Animate a 3d projection with depth
  cues.
}
\seealso{\code{\link{animate}} for options that apply to all animations}
\keyword{hplot}
\alias{display_depth}
\alias{animate_depth}
\arguments{
  \item{...}{other arguments passed on to \code{\link{animate}}}
}
\examples{animate_depth(flea[, 1:6])}
