% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-image.r
\name{display_image}
\alias{display_image}
\alias{animate_image}
\title{Image tour path animation.}
\usage{
display_image(xs, ys, ...)

animate_image(data, tour_path = grand_tour(1), ...)
}
\arguments{
\item{xs}{x limit that is used in making the size of the plot}

\item{ys}{y limit that is used in making the size of the plot}

\item{...}{other arguments passed on to \code{\link{animate}}}

\item{data}{matrix, or data frame containing numeric columns}

\item{tour_path}{tour path generator, defaults to 2d grand tour}
}
\description{
Animate a 1d tour path with an image plot.  This animation requires a
different input data structure, a 3d array.  The first two dimensions are
locations on a grid, and the 3rd dimension gives the observations to be
mixed with the tour.
}
\examples{
str(ozone)
animate_image(ozone)
}
\seealso{
\code{\link{animate}} for options that apply to all animations
}
\keyword{hplot}
