% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-pcp.r
\name{display_pcp}
\alias{display_pcp}
\alias{animate_pcp}
\title{Parallel coordinates tour path animation.}
\usage{
display_pcp(...)

animate_pcp(data, tour_path = grand_tour(3), ...)
}
\arguments{
\item{...}{other arguments passed on to \code{\link{animate}}}

\item{data}{matrix, or data frame containing numeric columns}

\item{tour_path}{tour path generator, defaults to 2d grand tour}
}
\description{
Animate a nD tour path with a parallel coordinates plot.
}
\details{
The lines show the observations, and the points, the values of the
projection matrix.
}
\examples{
animate_pcp(flea[, 1:6], grand_tour(3))
animate_pcp(flea[, 1:6], grand_tour(5))
}
\seealso{
\code{\link{animate}} for options that apply to all animations
}
\keyword{hplot}
