% Generated by roxygen2 (4.0.0): do not edit by hand
\name{display_scatmat}
\alias{animate_scatmat}
\alias{display_scatmat}
\title{Scatterplot matrix tour path animation.}
\usage{
display_scatmat(...)

animate_scatmat(data, tour_path = grand_tour(3), ...)
}
\arguments{
\item{...}{other arguments passed on to \code{\link{animate}}}

\item{data}{matrix, or data frame containing numeric columns}

\item{tour_path}{tour path generator, defaults to 2d grand tour}
}
\description{
Animate a nD tour path with a scatterplot matrix.
}
\details{
The lines show the observations, and the points, the values of the
projection matrix.
}
\examples{
animate_scatmat(flea[, 1:6], grand_tour(2))
animate_scatmat(flea[, 1:6], grand_tour(6))
}
\seealso{
\code{\link{animate}} for options that apply to all animations
}
\keyword{hplot}

