\name{display_stereo}
\title{Anaglpyh tour path animation.}
\usage{display_stereo(blue, red, ...)}

\description{
  Anaglpyh tour path animation.
}

\details{
  Uses red-blue anaglyphs to display a 3d tour path.
  You'll need some red- blue glasses to get much out of
  this displays!
}
\keyword{hplot}
\alias{display_stereo}
\alias{animate_stereo}
\arguments{
  \item{blue}{blue colour (for right eye)}
  \item{red}{red colour (for left eye)}
  \item{...}{other arguments passed on to \code{\link{animate}}}
}
\examples{animate_stereo(flea[, 1:6])}
