% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-stereo.r
\name{display_stereo}
\alias{display_stereo}
\alias{animate_stereo}
\title{Anaglpyh tour path animation.}
\usage{
display_stereo(blue, red, cex = 1, ...)

animate_stereo(
  data,
  tour_path = grand_tour(3),
  blue = rgb(0, 0.91, 0.89),
  red = rgb(0.98, 0.052, 0),
  ...
)
}
\arguments{
\item{blue}{blue colour (for right eye)}

\item{red}{red colour (for left eye)}

\item{cex}{size of the point to be plotted.  Defaults to 1.}

\item{...}{other arguments passed on to \code{\link{animate}}}

\item{data}{matrix, or data frame containing numeric columns}

\item{tour_path}{tour path generator, defaults to 2d grand tour}
}
\description{
Uses red-blue anaglyphs to display a 3d tour path.  You'll need some red-
blue glasses to get much out of this displays!
}
\examples{
animate_stereo(flea[, 1:6])
}
\keyword{hplot}
