% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-geodesic.r
\name{find_best_dir}
\alias{find_best_dir}
\title{Find the most promising direction to travel in.}
\usage{
find_best_dir(old, index, dist = 0.01, counter = 5, ...)
}
\arguments{
\item{old}{current projection}

\item{index}{interestingness index function}

\item{dist}{step size in radians, should be small}

\item{counter}{of random steps to take}
}
\description{
Starting from the current projection, pick \code{counter} random location
and take a small step towards and away from each location.  The most
promising direction has the highest value of the \code{index} function.
}
\keyword{internal}
\keyword{optimize}
