% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-frozen-geodesic.r
\name{find_best_frozen_dir}
\alias{find_best_frozen_dir}
\title{Find most promising direction in frozen space.}
\usage{
find_best_frozen_dir(old, frozen, index, dist = 0.01, counter = 5)
}
\description{
Find most promising direction in frozen space.
}
\keyword{internal}
