% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-frozen-geodesic.r
\name{find_frozen_path_peak}
\alias{find_frozen_path_peak}
\title{Find most highest peak along frozen geodesic.}
\usage{
find_frozen_path_peak(old, new, frozen, index, max_dist = pi/4)
}
\description{
Find most highest peak along frozen geodesic.
}
\keyword{internal}
