% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-geodesic.r
\name{find_path_peak}
\alias{find_path_peak}
\title{Find the most interesting projection along a geodesic.}
\usage{
find_path_peak(old, new, index, max_dist = pi/4, ...)
}
\arguments{
\item{old}{currention project}

\item{new}{projection that gives direction to travel in}

\item{index}{interestingness index function}

\item{max_dist}{maximum distance to travel along in radians}
}
\description{
Use \code{\link{optimize}} to find the most interesting projection amongst
all projections on a geodesic.  This method assumes that the function is
continuous with a single maximum, but seems to do ok even if there are
multiple maxima.
}
\keyword{internal}
\keyword{optimize}
