% Generated by roxygen2 (4.0.2): do not edit by hand
\name{holes}
\alias{holes}
\title{Holes index.}
\usage{
holes(mat)
}
\arguments{
\item{mat}{matrix being used}
}
\description{
Calculates the holes index. See Cook and Swayne (2007)
Interactive and Dynamic Graphics for Data Analysis for equations.
}
\keyword{hplot}

