% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is_orthonormal}
\alias{is_orthonormal}
\title{Test if a numeric matrix is orthonormal.}
\usage{
is_orthonormal(x, tol = 0.001)
}
\arguments{
\item{x}{numeric matrix}

\item{tol}{tolerance used to test floating point differences}
}
\description{
Test if a numeric matrix is orthonormal.
}
\keyword{algebra}
\keyword{internal}

