\name{is_orthonormal}
\alias{is_orthonormal}
\title{Test if a numeric matrix is orthonormal.}
\usage{is_orthonormal(x, tol=0.001)}

\description{
  Test if a numeric matrix is orthonormal.
}
\keyword{internal}
\keyword{algebra}
\arguments{
  \item{x}{numeric matrix}
  \item{tol}{tolerance used to test floating point differences}
}
