% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interesting-indices.r
\name{lda_pp}
\alias{lda_pp}
\title{LDA projection pursuit index.}
\usage{
lda_pp(cl)
}
\arguments{
\item{cl}{class to be used.  Such as "color"}
}
\description{
Calculate the LDA projection pursuit index.  See Cook and Swayne (2007)
Interactive and Dynamic Graphics for Data Analysis for equations.
}
\keyword{hplot}
