% Generated by roxygen2 (4.0.0): do not edit by hand
\name{normalise}
\alias{normalise}
\title{Normalise a numeric matrix.}
\usage{
normalise(x)
}
\arguments{
\item{x}{numeric matrix or vector}
}
\description{
Ensure that columns of a numeric matrix have norm 1
}
\keyword{algebra}
\keyword{internal}

