% Generated by roxygen2 (4.0.2): do not edit by hand
\name{orthonormalise}
\alias{orthonormalise}
\title{Orthonormalise using modified Gram-Schmidt process.}
\usage{
orthonormalise(x)
}
\arguments{
\item{x}{numeric matrix}
}
\description{
Orthonormalise using modified Gram-Schmidt process.
}
\keyword{algebra}
\keyword{internal}

