% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-algebra.r
\name{orthonormalise}
\alias{orthonormalise}
\title{Orthonormalise using modified Gram-Schmidt process.}
\usage{
orthonormalise(x)
}
\arguments{
\item{x}{numeric matrix}
}
\description{
Orthonormalise using modified Gram-Schmidt process.
}
\keyword{algebra}
\keyword{internal}
