% Generated by roxygen2 (4.0.2): do not edit by hand
\name{orthonormalise_by}
\alias{orthonormalise_by}
\title{Orthonnormalise one matrix by another.}
\usage{
orthonormalise_by(x, by)
}
\arguments{
\item{x}{numeric matrix}

\item{by}{numeric matrix, same size as x}
}
\description{
This ensures that each column in x is orthogonal to the corresponding
column in by.
}
\keyword{algebra}
\keyword{internal}

