% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-algebra.r
\name{orthonormalise_by}
\alias{orthonormalise_by}
\title{Orthonormalise one matrix by another.}
\usage{
orthonormalise_by(x, by)
}
\arguments{
\item{x}{numeric matrix}

\item{by}{numeric matrix, same size as x}
}
\value{
orthonormal numeric matrix
}
\description{
This ensures that each column in x is orthogonal to the corresponding
column in by.
}
\keyword{algebra}
\keyword{internal}
