% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interesting-indices.r
\name{pda_pp}
\alias{pda_pp}
\title{PDA projection pursuit index.}
\usage{
pda_pp(cl, lambda = 0.2)
}
\arguments{
\item{cl}{class to be used.  Such as "color"}

\item{lambda}{shrinkage parameter (0 = no shrinkage, 1 = full shrinkage)}
}
\description{
Calculate the PDA projection pursuit index.  See Lee and Cook (2009)
A Projection Pursuit Index for Large p, Small n Data
}
\keyword{hplot}
