% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.path_curve}
\alias{plot.path_curve}
\title{Plot history curves.}
\usage{
\method{plot}{path_curve}(x, ...)
}
\description{
The default plot method is a line plot with step on the x axis and
value on the y axis.  Each observation is drawn with a different line
line and the plot is facetted by variable.  This is rather similar in
spirit to a parallel coordinates plot or Andrews curves.
}
\details{
For alternative ways of plotting this data, see
 \code{\link{path_curves}}
}
\keyword{internal}

