% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path-index.r
\name{plot.path_index}
\alias{plot.path_index}
\title{Plot history index with ggplot2.}
\usage{
\method{plot}{path_index}(x, ...)
}
\description{
Plot history index with ggplot2.
}
\keyword{hplot}
\keyword{internal}
