\name{proj_dist}
\alias{proj_dist}
\title{Calculate the distance between two bases.}
\usage{
  proj_dist(x, y)
}
\arguments{
  \item{x}{projection matrix a}

  \item{y}{projection matrix b}
}
\description{
  Calculate the distance between two bases.
}
\details{
  Computes the Frobenius norm between two bases.  This is
  equals to the Euclidean norm of the vector of sines of
  principal angles between the two subspaces.
}
\keyword{algebra}

