% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-algebra.r
\name{proj_dist}
\alias{proj_dist}
\title{Calculate the distance between two bases.}
\usage{
proj_dist(x, y)
}
\arguments{
\item{x}{projection matrix a}

\item{y}{projection matrix b}
}
\description{
Computes the Frobenius norm between two bases, in radians.  This is
equals to the Euclidean norm of the vector of principal angles between
the two subspaces.
}
\keyword{algebra}
