% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{rescale}
\alias{rescale}
\title{Rescale a matrix or data frame}
\usage{
rescale(df)
}
\arguments{
\item{df}{data frame or matrix}
}
\description{
Standardise each column to have range [0, 1]
}
\keyword{manip}
