% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-better.r
\name{search_better}
\alias{search_better}
\title{Search for a better projection near the current projection.}
\usage{
search_better(current, alpha = 0.5, index, max.tries = Inf,
  method = "linear", cur_index = NA)
}
\description{
Search for a better projection near the current projection.
}
\keyword{internal}
