% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-better.r
\name{search_better_random}
\alias{search_better_random}
\title{Search for better projection, with stochastic component.}
\usage{
search_better_random(current, alpha = 0.5, index, max.tries = Inf,
  method = "linear", eps = 0.001, cur_index = NA)
}
\description{
Search for better projection, with stochastic component.
}
\keyword{internal}
