\name{search_better_random}
\alias{search_better_random}
\title{Search for better projection, with stochastic component.}
\usage{
  search_better_random(current, alpha = 0.5, index,
  max.tries = Inf, method = "linear", eps = 0.001)
}
\description{
  Search for better projection, with stochastic component.
}
\keyword{internal}

