% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_polish.R
\name{search_polish}
\alias{search_polish}
\title{Polish the searching after search_geodesic}
\usage{
search_polish(
  current,
  polish_alpha = 0.05,
  iter = 1000,
  nloop = 10,
  polish_cooling = 1,
  ...
)
}
\arguments{
\item{current}{starting projection}

\item{polish_alpha}{maximum distance to travel}

\item{iter}{number of sample basis to take in the sphere}

\item{nloop}{number of time the polish process get executed}
}
\description{
Polish the searching after search_geodesic
}
\keyword{optimize}
