% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_polish.R
\name{search_polish}
\alias{search_polish}
\title{Polish the searching after search_geodesic}
\usage{
search_polish(current, grid_gap = 0.002)
}
\arguments{
\item{current}{starting projection}

\item{alpha_polish}{maximum distance to travel (currently ignored)}

\item{index}{interestingness index function}

\item{max.tries}{maximum number of basis to search}

\item{method}{linear or geodesic nearby basis}
}
\description{
Polish the searching after search_geodesic
}
\keyword{optimize}
