% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_posse.R
\name{search_posse}
\alias{search_posse}
\title{Search for a better projection based on Poss, 1995}
\usage{
search_posse(
  current,
  alpha = 0.5,
  index,
  tries,
  max.tries = 300,
  cur_index = NA,
  ...
)
}
\arguments{
\item{current}{starting projection}

\item{alpha}{the angle used to search the target basis from the current basis}

\item{index}{index function}

\item{tries}{the counter of the outer loop of the opotimiser}

\item{max.tries}{maximum number of iteration before giving up}

\item{cur_index}{the index value of the current basis}

\item{...}{other arguments being passed into the \code{search_better()}}
}
\description{
Search for a better projection based on Poss, 1995
}
\keyword{optimize}
