% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sphere}
\alias{sphere}
\title{Sphere a matrix (or data frame) by transforming variables to
principal components.}
\usage{
sphere(df)
}
\arguments{
\item{df}{data frame or matrix}
}
\description{
Sphering is often useful in conjunction with the guided tour, as it
removes simpler patterns that may conceal more interesting findings.
}
\keyword{manip}

