\name{sphere}
\alias{sphere}
\title{Sphere a matrix (or data frame) by transforming variables to...}
\usage{sphere(df)}

\description{
  Sphere a matrix (or data frame) by transforming variables
  to principal components.
}

\details{
  Sphering is often useful in conjunction with the guided
  tour, as it removes simpler patterns that may conceal
  more interesting findings.
}
\keyword{manip}
\arguments{
  \item{df}{data frame or matrix}
}
