% Generated by roxygen2 (4.0.2): do not edit by hand
\name{step_angle}
\alias{step_angle}
\title{Step along an interpolated path by angle in radians.}
\usage{
step_angle(interp, angle)
}
\arguments{
\item{interp}{interpolated path}

\item{angle}{angle, in radians}
}
\description{
Step along an interpolated path by angle in radians.
}
\keyword{internal}

