\name{step_angle}
\alias{step_angle}
\title{Step along an interpolated path by angle in radians.}
\usage{step_angle(interp, angle)}

\description{
  Step along an interpolated path by angle in radians.
}
\keyword{internal}
\arguments{
  \item{interp}{interpolated path}
  \item{angle}{angle, in radians}
}
