% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodesic.r
\name{step_fraction}
\alias{step_fraction}
\title{Step along an interpolated path by fraction of path length.}
\usage{
step_fraction(interp, fraction)
}
\arguments{
\item{interp}{interpolated path}

\item{fraction}{fraction of distance between start and end planes}
}
\description{
Step along an interpolated path by fraction of path length.
}
\keyword{internal}
