% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section-pursuit.r
\name{weights_bincount_radial}
\alias{weights_bincount_radial}
\title{Computes weights for the rescaling of radial bin counts.}
\usage{
weights_bincount_radial(histo, p)
}
\arguments{
\item{histo}{input histogram (for bin definitions)}

\item{p}{number of variables}
}
\description{
Computes weights for the rescaling of radial bin counts.
}
\keyword{internal}
