% Generated by roxygen2 (4.0.0): do not edit by hand
\name{xnul}
\alias{nul}
\alias{xnul}
\title{A null function}
\usage{
nul(...)
}
\arguments{
\item{...}{all arguments to \code{...} are ignore}
}
\description{
This function does nothing, and is a useful default callback function
}
\keyword{internal}

