% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{xnul}
\alias{xnul}
\alias{nul}
\title{A null function}
\usage{
nul(...)
}
\arguments{
\item{...}{all arguments to \code{...} are ignore}
}
\description{
This function does nothing, and is a useful default callback function
}
\keyword{internal}
